Pulled a copy of the downloads links out of the  {% if ... %} section
The default.html page already had the following code:
 
     {% if site.show_downloads %}
        <a href="{{ site.github.zip_url }}" class="btn">Download .zip</a>
        <a href="{{ site.github.tar_url }}" class="btn">Download .tar.gz</a>
      {% endif %}

but no "downloads" appeared on the page. It may be that the variable "site.show_downloads" wasn't enabled somewhere in the GitHub control panel (or theme or somewhere else?). To see if the downloads would show, the two download lines were copied outside of the "if" block.